Silence compiler warnings about signed/unsigned int mixup
QShaderDescription::tessellationOutputVertexCount returns an unsigned
integer, producing a compiler warning when QCOMPARE'ing to an int
literal.

+++ tests/auto/qshaderbaker/tst_qshaderbaker.cpp
@@ -1082,7 +1082,7 @@ void tst_QShaderBaker::tessellationCompile()
     QCOMPARE(desc.outputBuiltinVariables()[1].type, QShaderDescription::TessLevelOuterBuiltin);
     QCOMPARE(desc.outputBuiltinVariables()[2].type, QShaderDescription::TessLevelInnerBuiltin);
 
-    QCOMPARE(desc.tessellationOutputVertexCount(), 3);
+    QCOMPARE(desc.tessellationOutputVertexCount(), uint(3));
 
     QCOMPARE(desc.outputVariables().size(), 3);
     for (const QShaderDescription::InOutVariable &v : desc.outputVariables()) {
